+++ app/src/main/AndroidManifest.xml
@@ -13,6 +13,13 @@
         android:theme="@style/Theme.EasySoccer"
         tools:targetApi="31">
         <activity
+            android:name=".NavigatorAdminActivity"
+            android:exported="false">
+            <meta-data
+                android:name="android.app.lib_name"
+                android:value="" />
+        </activity>
+        <activity
             android:name=".RegisterUserActivity"
             android:exported="false">
             <meta-data
