+++ app/src/main/java/com/wilman/easysoccer/ui/homeAdmin/HomeAdminFragment.kt
@@ -1,0 +1,38 @@
+package com.wilman.easysoccer.ui.homeAdmin
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.ViewModelProvider
+import com.wilman.easysoccer.databinding.FragmentHomeAdminBinding
+
+class HomeAdminFragment : Fragment() {
+
+    private var _binding: FragmentHomeAdminBinding? = null
+
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        val homeAdminViewModel =
+            ViewModelProvider(this)[HomeAdminViewModel::class.java]
+
+        _binding = FragmentHomeAdminBinding.inflate(inflater, container, false)
+        val root: View = binding.root
+
+
+        homeAdminViewModel.text.observe(viewLifecycleOwner) {
+        }
+        return root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
