+++ app/src/main/java/com/wilman/easysoccer/ui/homeAdmin/HomeAdminViewModel.kt
@@ -1,10 +1,10 @@
-package com.wilman.easysoccer.ui.home
+package com.wilman.easysoccer.ui.homeAdmin
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class HomeViewModel : ViewModel() {
+class HomeAdminViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
         value = "This is home Fragment"
