+++ app/src/main/java/com/wilman/easysoccer/ui/stadiums/StadiumsViewModel.kt
@@ -1,10 +1,10 @@
-package com.wilman.easysoccer.ui.dashboard
+package com.wilman.easysoccer.ui.stadiums
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class DashboardViewModel : ViewModel() {
+class StadiumsViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
         value = "This is dashboard Fragment"
