+++ app/src/main/res/layout/activity_navigator_admin.xml
@@ -16,7 +16,7 @@
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
-        app:menu="@menu/bottom_nav_menu" />
+        app:menu="@menu/bottom_nav_admin" />
 
     <fragment
         android:id="@+id/nav_host_fragment_activity_navigator_admin"
