+++ app/src/main/res/menu/bottom_nav_admin.xml
@@ -1,0 +1,22 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/navigation_home"
+        android:icon="@drawable/ic_home_black_24dp"
+        android:title="Inicio" />
+
+    <item
+        android:id="@+id/navigation_stadiums"
+        android:icon="@drawable/ic_dashboard_black_24dp"
+        android:title="Canchas" />
+
+    <item
+        android:id="@+id/navigation_centroDeportivo"
+        android:icon="@drawable/ic_notifications_black_24dp"
+        android:title="Centro Deportivo" />
+
+    <item
+        android:id="@+id/navigation_reservasAdmin"
+        android:icon="@drawable/ic_dashboard_black_24dp"
+        android:title="Reservas" />
+</menu>
