+++ app/src/main/res/values/strings.xml
@@ -2,7 +2,7 @@
     <string name="app_name">EasySoccer</string>
     <string name="title_activity_navigator_admin">NavigatorAdminActivity</string>
     <string name="title_home">Home</string>
-    <string name="title_dashboard">Dashboard</string>
+    <string name="title_stadiums">Canchas</string>
     <string name="title_notifications">Notifications</string>
     <string name="register_admin">Registro de Administradores</string>
     <string name="register_user">Registro de usuario</string>
